Adding Application

+++ src/main/java/alessiacotini/entities/Games.java
@@ -4,10 +4,10 @@ import java.time.Duration;
 import java.time.LocalDate;
 
 public class Games extends StartingForAll{
-    private byte numberOfPlayers;
+    private Integer numberOfPlayers;
     private final Duration duration;
 
-    public Games(String title, LocalDate yearOfRelease, double price, byte numberOfPlayers, int duration) {
+    public Games(String title, LocalDate yearOfRelease, double price, Integer numberOfPlayers, int duration) {
         super(title, yearOfRelease, price);
         this.duration = Duration.ofMinutes(duration);
         this.numberOfPlayers = (numberOfPlayers >= 2 && numberOfPlayers <= 10) ? numberOfPlayers : 2;
@@ -16,10 +16,10 @@ public class Games extends StartingForAll{
     public Integer getDuration() {
         return Math.toIntExact(duration.toMinutes());
     }
-    public byte getNumberOfPlayers() {
+    public Integer getNumberOfPlayers() {
         return numberOfPlayers;
     }
-    public void setNumberOfPlayers(byte numberOfPlayers) {
+    public void setNumberOfPlayers(Integer numberOfPlayers) {
         this.numberOfPlayers = numberOfPlayers;
     }
